abends and DDname freeing
I added an abend test so that I could verify I get an rc of 255 which
can then be tested for from a script.
I added skeleton code to free DDnames for all the DDnames allocated. I
have implemented PDS member and dataset. This was to fix a problem
where, specifically for ADRDSSU, I could not copy the file after mvscmd
completed and had to put a nasty sleep in. By explicitly dynfree()ing
the DDname, this seems to have solved the problem.